feat: structured output for llms
I've switched to using structured output for the llm responses to make
them easier to parse in the application code. llm has support for this
but the llm-mlx plugin does not, so I've switched to using Ollama and
llm-ollama. This will likely have a performance impact on inference as
the benefit of MLX is making better use of Apple Silicon.

Refs: #7